requirements.txt
asgiref==3.7.2
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django==filter-24.2
drf-yasg==1.21.7
ruff==0.2.1
